refactor(perf): split /ai-evolution + /ai-art into server shell + motion island
23rd + 24th in the pattern series.
